Web Scraper
Added Puppeteer and Cheerio to scrape Urban Dictionary's daily words which tend to be of a better stock than the randomizer. They are more meaningful, popular, and less insulting. With this ability, I added a catchup to the webscraper to get historical words as there seems to be 900 pages of daily, curated words.