Definitions & Solutions
Moved old solutions to a defintions collections. This allows me to capture all curated words to use a pool for the daily word. This was for two reasons: 1) some days are skipped for the curated word 2) some words or phrases have too large of a word to be a solution. Because of this, I need to not only select from and earlier day on a whim because of a nonexistent daily word, but also to skip words that are unuseable.

I am still collecting words that do not meet criteria as the criteria I set later may change at a later date. Maybe a word like HOTWIFING, could be made managable by giving the user the -ING as a hint, making the word HOTWIFE.